+++ graphql/resolvers/studentResolvers.js
@@ -1,9 +1,11 @@
-
 import Student from "../../models/student.js";
+import { authorize } from "../../middleware/authorize.js"
 
 export const studentResolvers = {
     Query: {
-        students: async () => {
+        students: async (context) => {
+            await authorize(context, ['Head admin', 'Branch admin', 'Librarian'])
+
             try {
                 return await Student.find();
             } catch (error) {
@@ -11,7 +13,9 @@ export const studentResolvers = {
                 throw new Error('Error fetching students');
             }
         },
-        student: async (_, { id }) => {
+        student: async (_, { id }, context) => {
+            await authorize(context, ['Head admin', 'Branch admin', 'Librarian'])
+
             try {
                 return await Student.findById(id);
             } catch (error) {
@@ -21,7 +25,9 @@ export const studentResolvers = {
         },
     },
     Mutation: {
-        addStudent: async (_, { studentId, name, age, gender, books_availed }) => {
+        addStudent: async (_, { studentId, name, age, gender, books_availed }, context) => {
+            await authorize(context, ['Head admin', 'Branch admin', 'Librarian'])
+
             try {
                 const newStudent = new Student({ studentId, name, age, gender, books_availed });
                 return await newStudent.save();
@@ -30,7 +36,9 @@ export const studentResolvers = {
                 throw new Error('Error adding new student');
             }
         },
-        updateStudent: async (_, { id, studentId, name, age, gender, books_availed }) => {
+        updateStudent: async (_, { id, studentId, name, age, gender, books_availed }, context) => {
+            await authorize(context, ['Head admin', 'Branch admin', 'Librarian'])
+
             try {
                 const updatedStudent = await Student.findByIdAndUpdate(
                     id,
@@ -43,7 +51,9 @@ export const studentResolvers = {
                 throw new Error('Error updating student');
             }
         },
-        deleteStudent: async (_, { id }) => {
+        deleteStudent: async (_, { id }, context) => {
+            await authorize(context, ['Head admin', 'Branch admin', 'Librarian'])
+
             try {
                 await Student.findByIdAndDelete(id);
                 return "Student has been deleted.";
